fidgeting around with the tuning process

+++ tune_parameters_for_benchmark.py
@@ -5,40 +5,45 @@ import skopt
 from skopt.space import Real, Categorical, Integer
 import dill
 import multiprocess as mp
-
+import os
 import sys
 
 if __name__ == '__main__':
     idx = int(sys.argv[1])
     name, recipe = list(benchmark_images.items())[idx]
+    print(name)
 
+    result_name = name+'_starfinder_opt.pkl'
+    if not os.path.exists(result_name):
+        starfinder_obj = make_starfinder_objective(recipe, name)
+        starfinder_dims = [
+            Real(-6., 3, name='threshold'),
+            Real(2., 8, name='fwhm'),
+            Real(2., 5.5, name='sigma_radius'),
+            Real(-15., 0., name='roundlo'),
+            Real(0., 15., name='roundhi'),
+            Real(-15, 0., name='sharplo'),
+            Real(0., 10., name='sharphi')
+        ]
 
-    starfinder_obj = make_starfinder_objective(recipe, name)
-    starfinder_dims = [
-        Real(-6., 3, name='threshold'),
-        Real(2., 8, name='fwhm'),
-        Real(2., 5.5, name='sigma_radius'),
-        Real(-15., 0., name='roundlo'),
-        Real(0., 15., name='roundhi'),
-        Real(-15, 0., name='sharplo'),
-        Real(0., 10., name='sharphi')
-    ]
+        starfinder_optimizer = skopt.Optimizer(
+            dimensions=starfinder_dims,
+            n_jobs=12,
+            random_state=1,
+            base_estimator='RF',
+            n_initial_points=1000,
+            initial_point_generator='random'
+        )
 
-    starfinder_optimizer = skopt.Optimizer(
-        dimensions=starfinder_dims,
-        n_jobs=12,
-        random_state=1,
-        base_estimator='RF',
-        n_initial_points=1000,
-        initial_point_generator='random'
-    )
+        starfinder_result = run_optimizer(starfinder_optimizer, starfinder_obj, n_evaluations=1400, n_processes=12)
 
-    starfinder_result = run_optimizer(starfinder_optimizer, starfinder_obj, n_evaluations=1400, n_processes=12)
-
-    with open(name+'_starfinder_opt.pkl', 'wb') as f:
-        dill.dump(starfinder_result, f)
-    with open(name+'starfinder_params.txt', 'w') as f:
-        f.write(str(list(zip(starfinder_result.space.dimension_names, starfinder_result.x))))
+        with open(name+'_starfinder_opt.pkl', 'wb') as f:
+            dill.dump(starfinder_result, f)
+        with open(name+'starfinder_params.txt', 'w') as f:
+            f.write(str(list(zip(starfinder_result.space.dimension_names, starfinder_result.x))))
+    else:
+        with open(result_name, 'rb') as f:
+            starfinder_result = dill.load(f)
 
     x = starfinder_result.x
     config = Config()
@@ -59,7 +64,8 @@ if __name__ == '__main__':
         Integer(1, 10, name='fitshape_half'),
         Real(0.30, 3., name='sigma'),
         Categorical([4, 5, 7, 10], name='iters'),
-        Categorical([1, 2, 4], name='oversampling')
+        Categorical([1, 2, 4], name='oversampling'),
+        Real(0.5, 2, name='separation factor')
     ]
 
     epsf_optimizer = skopt.Optimizer(
@@ -67,10 +73,10 @@ if __name__ == '__main__':
         n_jobs=12,
         random_state=1,
         base_estimator='RF',
-        n_initial_points=1000,
+        n_initial_points=50,
         initial_point_generator='random'
     )
-    epsf_result = run_optimizer(epsf_optimizer, epsf_obj, n_evaluations=1400, n_processes=12)
+    epsf_result = run_optimizer(epsf_optimizer, epsf_obj, n_evaluations=200, n_processes=9)
 
     with open(name+'_epsf_opt.pkl', 'wb') as f:
         dill.dump(epsf_result, f)
